code/integrations_service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pika